Added some of my solutions
I have added some of the questions which I had solved .
