+++ Hackerearth/Practice/Basic Programming/Palindromic String.py
@@ -1,0 +1,9 @@
+#LINK - https://www.hackerearth.com/practice/basic-programming/input-output/basics-of-input-output/practice-problems/algorithm/palindrome-check-2/
+
+s = input()
+r = s[::-1]
+
+if ( s==r ) :
+	print("YES")
+else :
+	print("NO")
