+++ Hackerearth/Practice/Basic Programming/Read from STDIN, Write to STDOUT.c
@@ -1,0 +1,12 @@
+//LINK - https://www.hackerearth.com/practice/basic-programming/input-output/basics-of-input-output/practice-problems/algorithm/io-exercise/
+
+#include <stdio.h>
+
+int main()
+{
+   int a;
+    char s[15];
+    scanf("%d%s",&a,&s);
+    printf("%d\n%s",2*a,s);
+    return 0;
+}
