+++ Hackerearth/Practice/Basic Programming/Read from STDIN, Write to STDOUT.py
@@ -1,0 +1,6 @@
+#LINK1 - https://www.hackerearth.com/practice/basic-programming/input-output/basics-of-input-output/practice-problems/algorithm/io-exercise/
+
+n=int(input())
+s=input()
+print(2*n)
+print(s)
